+++ src/main/java/com/gofobao/framework/system/vo/response/IndexStatistics.java
@@ -27,10 +27,10 @@ public class IndexStatistics {
     private Long borrowTotal;
 
     @ApiModelProperty("今日成交额")
-    private Integer todayDueTotal;
+    private Long todayDueTotal;
 
     @ApiModelProperty("昨日成交额")
-    private Integer yesterdayDueTotal;
+    private Long yesterdayDueTotal;
 
     @ApiModelProperty("起头金额(元)")
     private Integer startMoney;
